Add: repo

+++ server/apppyme/src/main/java/com/sv/apppyme/entities/DetalleFac.java
@@ -11,8 +11,8 @@ public class DetalleFac {
 	private CabeceraFac cabeceraFac;
 	private Producto producto;
 	private int cantidad;
-	private double precioUnitario;
-	private double subTotal;
+	private float precioUnitario;
+	private float subTotal;
 
 	public DetalleFac() {
 		super();
